[fix](csv reader) fix data loss when concurrency read using multi char line delimiter (#53374) (#4199)
pick [(#53374)](https://github.com/apache/doris/pull/53374)

Multiple concurrent split file locations will be determined in plan
phase, if the split point happens to be in the middle of the multi char
line delimiter:

- The previous concurrent will read the complete row1 and read a little
more to read the line delimiter.
- The latter concurrency will start reading from half of the multi char
line delimiter, and row2 is the first line of this concurrency, but the
first line in the middle range is always discarded, so row2 will be
lost.

## Proposed changes

Issue Number: close #xxx

<!--Describe your changes.-->